Update VAT type to decimal, fix due date validation
Changed VAT (ALV) fields from float to decimal for improved precision and updated default to 25.5. Adjusted all related calculations and database reads to use decimal.

+++ Tietokanta.cs
@@ -347,7 +347,7 @@ namespace Harjoitustyö
                                 Määrä = r.GetInt32("määrä"),
                                 Yksikkö = r.GetString("yksikkö"),
                                 A_Hinta = r.GetDecimal("a_hinta"),
-                                ALV = r.GetFloat("alv")
+                                ALV = r.GetDecimal("alv")
                             });
                         }
                     }
@@ -393,7 +393,7 @@ namespace Harjoitustyö
                                 Määrä = r.GetInt32("määrä"),
                                 Yksikkö = r.GetString("yksikkö"),
                                 A_Hinta = r.GetDecimal("a_hinta"),
-                                ALV = r.GetFloat("alv")
+                                ALV = r.GetDecimal("alv")
                             });
                         }
                     }
